+++ src/ralph/api/__init__.py
@@ -4,8 +4,10 @@ Main module for Ralph's LRS API.
 from fastapi import Depends, FastAPI
 
 from .auth import AuthenticatedUser, authenticated_user
+from .routers import statements
 
 app = FastAPI()
+app.include_router(statements.router)
 
 
 @app.get("/whoami")